Pull out lineage functions into lineage library
