+++ lib/utils.js
@@ -487,45 +487,6 @@ module.exports = {
       return phoneUtil.isNumberMatch(gw, from) >= 3;
     }
     return false;
-  },
-  hydrateDoc: (db, id, callback) => {
-    db.medic.view('medic-client', 'docs_by_id_lineage', {
-      startkey: [ id ],
-      endkey: [ id, {} ],
-      include_docs: true
-    }, (err, lineage) => {
-      if (err) {
-        return callback(err);
-      }
-      const contactIds = lineage.rows
-        .map(row => row.doc && row.doc.contact && row.doc.contact._id)
-        .filter(id => !!id);
-      db.medic.fetch({ keys: contactIds }, (err, contacts) => {
-        if (err) {
-          return callback(err);
-        }
-        contacts.rows.forEach(contactRow => {
-          const stub = lineage.rows.find(lineageRow => {
-            const id = lineageRow.doc &&
-                       lineageRow.doc.contact &&
-                       lineageRow.doc.contact._id;
-            return id === (contactRow.doc && contactRow.doc._id);
-          });
-          if (stub && stub.doc) {
-            stub.doc.contact = contactRow.doc;
-          }
-        });
-        const resultRow = lineage.rows.shift();
-        const result = resultRow && resultRow.doc;
-        let current = result;
-        while (current) {
-          const row = lineage.rows.shift();
-          current.parent = row && row.doc;
-          current = current.parent;
-        }
-        callback(null, result);
-      });
-    });
   }
 
 };
