+++ transitions/index.js
@@ -7,7 +7,7 @@ const _ = require('underscore'),
       follow = require('follow'),
       async = require('async'),
       utils = require('../lib/utils'),
-      transitionUtils = require('./utils'),
+      lineage = require('../lib/lineage'),
       logger = require('../lib/logger'),
       config = require('../config'),
       db = require('../db'),
@@ -76,7 +76,7 @@ const processChange = (change, callback) => {
   if (!change) {
     return callback();
   }
-  utils.hydrateDoc(db, change.id, (err, doc) => {
+  lineage.fetchHydratedDoc(change.id, (err, doc) => {
     if (err) {
       logger.error(`transitions: fetch failed for ${change.id} (${err})`);
       return callback();
@@ -205,13 +205,7 @@ const finalize = (options, callback) => {
   logger.debug(
     `calling audit.saveDoc on doc ${change.id} seq ${change.seq}`);
 
-  change.doc.parent = transitionUtils.extractLineage(change.doc.parent);
-  if (change.doc.contact && change.doc.contact._id) {
-    change.doc.contact = {
-      _id: change.doc.contact._id,
-      parent: transitionUtils.extractLineage(change.doc.contact.parent)
-    };
-  }
+  lineage.minify(change.doc);
   audit.saveDoc(change.doc, err => {
     // todo: how to handle a failed save? for now just
     // waiting until next change and try again.
