+++ transitions/utils.js
@@ -56,18 +56,5 @@ module.exports = {
       doc.patient_id = patientId;
       callback();
     }).catch(callback);
-  },
-  extractLineage: contact => {
-    if (!contact) {
-      return contact;
-    }
-    const result = { _id: contact._id };
-    let minified = result;
-    while(contact.parent) {
-      minified.parent = { _id: contact.parent._id };
-      minified = minified.parent;
-      contact = contact.parent;
-    }
-    return result;
   }
 };
